Only call .toString() on sass files
Calling .toString() is a fairly expensive operation, which allocates
a new string, and copies all of the contents of the file Buffer into
that string. This is fairly fast for small files or small sites, but
quickly starts to take seconds of the total build time on medium
sized sites (in my case, roughly 2 seconds of the overall build was
this call to toString()!)

This change makes it so we only call toString() if we actually want
the result - i.e. if it as a sass file. Thus, we don't call it on
markdown files, image files, video files, etc.